feat(admin): searchable host multi-select in session form
Host checkboxes don't scale to hundreds of guests. Reuse the site's
SelectHosts combobox, moved to app/select-hosts.tsx and generalized to
any { id, name } items.
